完成Tag功能、Staff功能、Animation dao；完成图片上传机制；完成一个json parse替代@RequestBody。

+++ src/main/kotlin/com/heerkirov/animation/model/Animation.kt
@@ -4,7 +4,7 @@ import com.heerkirov.animation.enums.*
 import java.time.LocalDate
 import java.time.LocalDateTime
 
-data class Animation(val id: Long,
+data class Animation(val id: Int,
                      val title: String,
                      val originTitle: String?,
                      val otherTitle: String?,
